+++ docs/components/modal-template.vue
@@ -16,7 +16,7 @@
       Last result: <strong>{{ lastResult }}</strong>
     </v-alert>
 
-    <df-modal v-model="isOpen" :title="title" :actions="actions" closable icon="mdi-login">
+    <df-modal v-model="isOpen" :title="title" :actions="actions" icon="mdi-login">
       <template #body>
         <v-text-field v-model="username" label="Username" autofocus />
         <v-text-field v-model="password" label="Password" type="password" />